Make CFs less polymorphic: represent all actions as strings
The HOL type system is not really helping us here. Since we want
morphisms between CFs with different sets of actions, it is better to
embed all the actions in a common universe.

Strings are nice for examples. The only restriction is that they are
countable, but that should be fine for most purposes.