Make cf non-polymorphic
Polymorphism was in the type of world outcomes. Now these are also just
strings. This is because we want cf1 (e.g.) to have the agent/env
choices literally be the outcomes.

For some reason this change slowed down the simplifier a lot, so have
also fixed/improved some proofs in this commit.